+++ CHANGELOG.md
@@ -82,6 +82,8 @@ The major version bump is due to dropped support for PHP `7.2` and a several bre
 
 --------------------------------------------------------
 
+Update using ```composer update```.
+
 ### 6.3.10, <small>(2020-07-09)</small>
 
 #### Bugfixes